requirements.txt
coqui-tts
faster-whisper
pyaudio
numpy
google-generativeai
wave